requirements.txt
ascon
numpy
pillow
pycryptodome
python-dotenv
streamlit
